+++ configure/etc/conf/plugins.properties
@@ -30,6 +30,7 @@ datacap-plugin-monetdb=plugin/datacap-plugin-monetdb/pom.xml
 datacap-plugin-ydb=plugin/datacap-plugin-ydb/pom.xml
 datacap-plugin-neo4j=plugin/datacap-plugin-neo4j/pom.xml
 datacap-plugin-oceanbase=plugin/datacap-plugin-oceanbase/pom.xml
+datacap-plugin-oracle=plugin/datacap-plugin-oracle/pom.xml
 datacap-convert-txt=convert/datacap-convert-txt/pom.xml
 datacap-convert-json=convert/datacap-convert-json/pom.xml
 datacap-convert-xml=convert/datacap-convert-xml/pom.xml
