--- plugin/datacap-jdbc-oracle/src/main/java/io/edurt/datacap/plugin/jdbc/oracle/OraclePlugin.java
@@ -1,86 +1,0 @@
-package io.edurt.datacap.plugin.jdbc.oracle;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import io.edurt.datacap.spi.PluginService;
-import io.edurt.datacap.spi.PluginType;
-import io.edurt.datacap.spi.adapter.JdbcAdapter;
-import io.edurt.datacap.spi.connection.JdbcConfigure;
-import io.edurt.datacap.spi.connection.JdbcConnection;
-import io.edurt.datacap.spi.model.Configure;
-import io.edurt.datacap.spi.model.Response;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.beanutils.BeanUtils;
-import org.apache.commons.lang3.ObjectUtils;
-
-@Slf4j
-@SuppressFBWarnings(value = {"EI_EXPOSE_REP"})
-public class OraclePlugin
-        implements PluginService
-{
-    private JdbcConfigure jdbcConfigure;
-    private JdbcConnection connection;
-    private Response response;
-
-    @Override
-    public String validator()
-    {
-        return "SELECT version FROM PRODUCT_COMPONENT_VERSION\n" +
-                "WHERE product LIKE 'Oracle Database%'";
-    }
-
-    @Override
-    public String name()
-    {
-        return "Oracle";
-    }
-
-    @Override
-    public String description()
-    {
-        return "Integrate Oracle data sources";
-    }
-
-    @Override
-    public PluginType type()
-    {
-        return PluginType.JDBC;
-    }
-
-    @Override
-    public void connect(Configure configure)
-    {
-        try {
-            this.response = new Response();
-            this.jdbcConfigure = new JdbcConfigure();
-            BeanUtils.copyProperties(this.jdbcConfigure, configure);
-            this.jdbcConfigure.setJdbcDriver("oracle.jdbc.OracleDriver");
-            this.jdbcConfigure.setJdbcType("oracle:thin");
-            this.connection = new OracleConnection(this.jdbcConfigure, this.response);
-        }
-        catch (Exception ex) {
-            this.response.setIsConnected(Boolean.FALSE);
-            this.response.setMessage(ex.getMessage());
-        }
-    }
-
-    @Override
-    public Response execute(String content)
-    {
-        if (ObjectUtils.isNotEmpty(this.connection)) {
-            log.info("Execute oracle plugin logic started");
-            this.response = this.connection.getResponse();
-            JdbcAdapter processor = new OracleAdapter(this.connection);
-            this.response = processor.handlerExecute(content);
-            log.info("Execute oracle plugin logic end");
-        }
-        return this.response;
-    }
-
-    @Override
-    public void destroy()
-    {
-        if (ObjectUtils.isNotEmpty(this.connection)) {
-            this.connection.destroy();
-        }
-    }
-}
