+++ plugin/datacap-plugin-oracle/pom.xml
@@ -10,13 +10,11 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>datacap-jdbc-oracle</artifactId>
-    <description>DataCap - Oracle</description>
+    <artifactId>datacap-plugin-oracle</artifactId>
+    <description>DataCap - Plugin - Oracle</description>
 
     <properties>
         <oracle.version>23.3.0.23.09</oracle.version>
-        <oracle-xe.version>1.20.1</oracle-xe.version>
-        <plugin.name>jdbc-oracle</plugin.name>
     </properties>
 
     <dependencies>
@@ -26,19 +24,9 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>commons-beanutils</groupId>
-            <artifactId>commons-beanutils</artifactId>
-        </dependency>
-        <dependency>
             <groupId>com.oracle.database.jdbc</groupId>
             <artifactId>ojdbc8</artifactId>
             <version>${oracle.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.testcontainers</groupId>
-            <artifactId>oracle-xe</artifactId>
-            <version>${oracle-xe.version}</version>
-            <scope>test</scope>
         </dependency>
     </dependencies>
 
@@ -61,6 +49,24 @@
                         <goals>
                             <goal>single</goal>
                         </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-dependency-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>copy-dependencies</id>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>copy-dependencies</goal>
+                        </goals>
+                        <configuration>
+                            <outputDirectory>${project.build.directory}/dependencies</outputDirectory>
+                            <excludeScope>provided</excludeScope>
+                            <includeScope>runtime</includeScope>
+                        </configuration>
                     </execution>
                 </executions>
             </plugin>
