+++ plugin/datacap-plugin-oracle/src/main/java/io/edurt/datacap/plugin/OraclePlugin.java
@@ -1,0 +1,5 @@
+package io.edurt.datacap.plugin;
+
+public class OraclePlugin
+        extends Plugin
+{}
