+++ pom.xml
@@ -58,7 +58,7 @@
         <module>plugin/datacap-plugin-ydb</module>
         <module>plugin/datacap-plugin-neo4j</module>
         <module>plugin/datacap-plugin-oceanbase</module>
-        <!--        <module>plugin/datacap-jdbc-oracle</module>-->
+        <module>plugin/datacap-plugin-oracle</module>
         <!--        <module>plugin/datacap-jdbc-phoenix</module>-->
         <!--        <module>plugin/datacap-jdbc-postgresql</module>-->
         <!--        <module>plugin/datacap-jdbc-presto</module>-->
